bug: remove localeProjection and storeProjection from priceSelection and add to ProductProjectionSearch
