+++ commercetools-models/src/main/java/io/sphere/sdk/products/search/PriceSelectionBuilder.java
@@ -23,10 +23,6 @@ public final class PriceSelectionBuilder extends Base implements Builder<PriceSe
     private String priceCustomerGroup;
     @Nullable
     private String priceChannel;
-    @Nullable
-    private String storeProjection;
-    @Nullable
-    private List <String> localeProjection = new ArrayList<>();
 
     private PriceSelectionBuilder(@Nonnull final String priceCurrency) {
         this.priceCurrency = priceCurrency;
@@ -44,9 +40,7 @@ public final class PriceSelectionBuilder extends Base implements Builder<PriceSe
         return ofCurrencyCode(template.getPriceCurrency())
                 .priceCountryCode(template.getPriceCountry())
                 .priceCustomerGroupId(template.getPriceCustomerGroup())
-                .priceChannelId(template.getPriceChannel())
-                .storeProjection(template.getStoreProjection())
-                .localeProjection(template.getLocaleProjection());
+                .priceChannelId(template.getPriceChannel());
     }
 
     public PriceSelectionBuilder priceCurrencyCode(@Nonnull final String priceCurrencyCode) {
@@ -88,30 +82,6 @@ public final class PriceSelectionBuilder extends Base implements Builder<PriceSe
         return this;
     }
 
-    public PriceSelectionBuilder storeProjection(@Nullable final String storeProjection) {
-        this.storeProjection = storeProjection;
-        return this;
-    }
-
-    public PriceSelectionBuilder localeProjection(@Nullable final List<String> localeProjection) {
-        this.localeProjection = localeProjection;
-        return this;
-    }
-
-    public PriceSelectionBuilder localeProjection(@Nullable final String localeProjection) {
-        this.localeProjection = new ArrayList<>();
-        this.localeProjection.add(localeProjection);
-        return this;
-    }
-
-    public PriceSelectionBuilder plusLocaleProjection(@Nullable final String localeProjection) {
-        if (this.localeProjection == null) {
-            this.localeProjection = new ArrayList<>();
-        }
-
-        this.localeProjection.add(localeProjection);
-        return this;
-    }
 
     @Nullable
     public String getPriceChannel() {
@@ -133,18 +103,8 @@ public final class PriceSelectionBuilder extends Base implements Builder<PriceSe
         return priceCustomerGroup;
     }
 
-    @Nullable
-    public String getStoreProjection() {
-        return storeProjection;
-    }
-
-    @Nullable
-    public List<String> getLocaleProjection() {
-        return localeProjection;
-    }
-
     @Override
     public PriceSelectionDsl build() {
-        return new PriceSelectionDsl(priceCurrency, priceCountry, priceCustomerGroup, priceChannel, storeProjection, localeProjection);
+        return new PriceSelectionDsl(priceCurrency, priceCountry, priceCustomerGroup, priceChannel);
     }
 }
