+++ commercetools-models/src/main/java/io/sphere/sdk/products/search/PriceSelectionDsl.java
@@ -20,18 +20,12 @@ public final class PriceSelectionDsl extends Base implements PriceSelection {
     private final String priceCustomerGroup;
     @Nullable
     private final String priceChannel;
-    @Nullable
-    private final String storeProjection;
-    @Nullable
-    private final List<String> localeProjection;
 
-    PriceSelectionDsl(@Nonnull final String priceCurrency, final String priceCountry, final String priceCustomerGroup, final String priceChannel, final String storeProjection, final List<String> localeProjection) {
+    PriceSelectionDsl(@Nonnull final String priceCurrency, final String priceCountry, final String priceCustomerGroup, final String priceChannel) {
         this.priceChannel = priceChannel;
         this.priceCurrency = priceCurrency;
         this.priceCountry = priceCountry;
         this.priceCustomerGroup = priceCustomerGroup;
-        this.storeProjection = storeProjection;
-        this.localeProjection = localeProjection;
     }
 
     public static PriceSelectionDsl ofCurrencyCode(final String currencyCode) {
@@ -78,18 +72,6 @@ public final class PriceSelectionDsl extends Base implements PriceSelection {
         return newBuilder().priceChannel(priceChannel).build();
     }
 
-    public PriceSelectionDsl withStoreProjection(@Nullable final String storeProjection) {
-        return newBuilder().storeProjection(storeProjection).build();
-    }
-
-    public PriceSelectionDsl withLocaleProjection(@Nullable final List<String> localeProjection) {
-        return newBuilder().localeProjection(localeProjection).build();
-    }
-
-    public PriceSelectionDsl plusLocaleProjection(@Nullable final String localeProjection) {
-        return newBuilder().plusLocaleProjection(localeProjection).build();
-    }
-
     @Override
     @Nullable
     public String getPriceChannel() {
@@ -112,17 +94,5 @@ public final class PriceSelectionDsl extends Base implements PriceSelection {
     @Nullable
     public String getPriceCustomerGroup() {
         return priceCustomerGroup;
-    }
-
-    @Override
-    @Nullable
-    public String getStoreProjection() {
-        return storeProjection;
-    }
-
-    @Override
-    @Nullable
-    public List<String> getLocaleProjection() {
-        return localeProjection;
     }
 }
