+++ commercetools-models/src/main/java/io/sphere/sdk/products/search/PriceSelectionQueryParameters.java
@@ -15,9 +15,7 @@ public final class PriceSelectionQueryParameters extends Base {
     public static final String PRICE_COUNTRY = "priceCountry";
     public static final String PRICE_CUSTOMER_GROUP = "priceCustomerGroup";
     public static final String PRICE_CHANNEL = "priceChannel";
-    public static final String STORE_PROJECTION = "storeProjection";
-    public static final String LOCALE_PROJECTION = "localeProjection";
-    public static final List<String> ALL_PARAMETERS = asList(PRICE_CURRENCY, PRICE_COUNTRY, PRICE_CUSTOMER_GROUP, PRICE_CHANNEL, STORE_PROJECTION, LOCALE_PROJECTION);
+    public static final List<String> ALL_PARAMETERS = asList(PRICE_CURRENCY, PRICE_COUNTRY, PRICE_CUSTOMER_GROUP, PRICE_CHANNEL);
 
     private PriceSelectionQueryParameters() {
     }
@@ -41,8 +39,6 @@ public final class PriceSelectionQueryParameters extends Base {
             addParamIfNotNull(resultingParameters, PRICE_COUNTRY, priceSelection.getPriceCountry());
             addParamIfNotNull(resultingParameters, PRICE_CUSTOMER_GROUP, priceSelection.getPriceCustomerGroup());
             addParamIfNotNull(resultingParameters, PRICE_CHANNEL, priceSelection.getPriceChannel());
-            addParamIfNotNull(resultingParameters, STORE_PROJECTION, priceSelection.getStoreProjection());
-            priceSelection.getLocaleProjection().forEach(parameter -> addParamIfNotNull(resultingParameters, LOCALE_PROJECTION, parameter));
         }
         return resultingParameters;
     }
@@ -64,8 +60,6 @@ public final class PriceSelectionQueryParameters extends Base {
                 .priceCountryCode(map.get(PRICE_COUNTRY))
                 .priceCustomerGroupId(map.get(PRICE_CUSTOMER_GROUP))
                 .priceChannelId(map.get(PRICE_CHANNEL))
-                .storeProjection(map.get(STORE_PROJECTION))
-                .localeProjection(Collections.singletonList(map.get(LOCALE_PROJECTION)))
                 .build();
     }
